Make messages look pretty
All messages will now be embeds if possible. The specific format of these messages can be found in embed.py, and the configure command will be accompanied by a reaction in the case that the bot does not have sufficient permissions to send a message in the guild. There is a deliberate choice to avoid having the bot send a DM in this case, since this restriction is likely to be known and to be a conscious decision to the admins of a server.