Add in rudimentary server verification
Now the bot requires its integration's role to have 0x2F3136 as its color. This should not impact any aesthetic organization that a server may use, as they can override the color with another role. The only scheme where this is not the case is if the color is 0x000000, but you can simply override the color with something similar enough to that, and white mode is so niche that I don't see the need to accomodate such users in such a scheme.