Enable code block uploads
Now the bot can use code put in code blocks. I've also refactored the code to use a custom Converter, but that's a minor detail and this is a personal project. Ain't nobody got time for that. Why use 16 commits when I can use 1 and just stuff everything in the body?